修复横屏适配bug、修复双击缩放动画

+++ SDPhotoBrowser/SDPhotoBrowser/SDPhotoBrowser.h
@@ -27,7 +27,7 @@
 @interface SDPhotoBrowser : UIView <UIScrollViewDelegate>
 
 @property (nonatomic, weak) UIView *sourceImagesContainerView;
-@property (nonatomic, assign) int currentImageIndex;
+@property (nonatomic, assign) NSInteger currentImageIndex;
 @property (nonatomic, assign) NSInteger imageCount;
 
 @property (nonatomic, weak) id<SDPhotoBrowserDelegate> delegate;
